+++ packages/kit/src/core/sync/create_manifest_data/index.js
@@ -35,7 +35,7 @@ export default function create_manifest_data({
 /**
  * @param {import('types').ValidatedConfig} config
  */
-function create_assets(config) {
+export function create_assets(config) {
 	return list_files(config.kit.files.assets).map((file) => ({
 		file,
 		size: fs.statSync(path.resolve(config.kit.files.assets, file)).size,
@@ -289,7 +289,12 @@ function create_routes_and_nodes(cwd, config, fallback) {
 	// and smaller indexes take fewer bytes. also, this guarantees that
 	// the default error/layout are 0/1
 	for (const route of routes) {
-		if (route.layout) nodes.push(route.layout);
+		if (route.layout) {
+			if (!route.layout?.component) {
+				route.layout.component = posixify(path.relative(cwd, `${fallback}/layout.svelte`));
+			}
+			nodes.push(route.layout);
+		}
 		if (route.error) nodes.push(route.error);
 	}